Fix emit of epilog codes
We may not have recorded any restore codes
for the registers.  Those are only done if we
stored different registers than we restore.

If that's the case then we need to replay
the DW_CFA_offset codes, but we'll have
to reverse them to match the order that
they are actually emitted.

We can also drop DW_CFA_mono_fp_alloc_info_win64_arm64,
it is not used.