Fix Windows ARM64 crash on Thread.Abort
The R18 register is used for the Thread Environment Block,
which is OS managed and should never change.

In addition calls to GetThreadContext will not return it as part of the
CONTEXT_INTEGER (documented in the windows headers).  You can
specifically ask for it by passing CONTEXT_ARM64_X18, but that would
need to be done consistently.  Since that could be forgotten and there's
no reason to save/restore it anyway it's better to just never touch it.