Avoiding a memory leak in the Timer class SchedulerLoop
Rather than calling the Sort overload that accepts the IComparer
interface, we will call the overload that accepts the Comparison
delegate.  Both overloads end up in the same place eventually.

There is a memory leak in the Mono runtime with the ldvirtftn opcode
in a generic method. The opcode targets a generic interfaces method on a
struct instance. Each time ldvirtftn is called an unbox trampoline is
created. The ldvirtftn leak will be filed separately upstream but this
simple change avoids the bug for this scenario.