Fix invalid check on value types (UUM-79115).
The previously inserted null check on 'this' assumed, at least on amd64,
that the pointer could be dereferenced as a pointer sized value. This is
not true for value types which can be smaller than pointer sized,
especially if the memory adjacent to the value type is protected.

Rather than improving the null check, we can avoid it entirely for value
type methods as they should never have a NULL 'this' pointer.