Fix/math operation error uum 85488 (#2086) (#2087)
* [mono] Fix a few corner case overflow operations (#58114)

* [mono][jit] Fix invalid uses of SHL instead of MUL optimization

mono_is_power_of_two is meant to be used only on unsigned numbers. In some cases we pass a signed value instead. This is typically not a problem because negative numbers are not detected as a power of two, except for the special number -2147483648, which is coded as 0x80000000, therefore a power of two.

* Enable tests

* [interp] Fix a few overflow conversions

Floating point numbers are truncated towards 0 when converted to integers, therefore we need to increase the overflow range. For example, (int8)-128.5 = -128, without overflows, even if -128.5 is smaller than the minimum integer.

Stop relying on undefined behavior by casting from floating point to integral without range checks.

* [interp] Extract some conversion logic in mono-math.h

Later we should reuse these methods also with jit

* Avoid transformation from multiplication to left shift in case of 64 bit value (#71189)

* Cherry-picked 2 commits from the upstream repo to fix issue with math operations, this change removes a dependency on another commit.

---------

Co-authored-by: Vlad Brezae <brezaevlad@gmail.com>
Co-authored-by: Alhad Deshpande <97085048+alhad-deshpande@users.noreply.github.com>