Debugger: Prevent Abort on Invalid Virtual Method
If the debugger is asked to invoke an interface or virtual
method on a class the mono_object_get_virtual_method_internal
will assert & abort (or crash).

mono_object_get_virtual_method_internal assumes that
it is being asked to invoke a virtual method on an instance
that implements it.  If that's not true it will assert & abort or
access invalid memory if assertions are disabled.

To protect against this we check if the this_arg can be cast to
the interface or base class type.  If it can then we should be
safe to call mono_object_get_virtual_method_internal.