Protect calls to fork with a gc_lock on MacOS
On MacOS running under Rosetta there is a hang if you are calling fork
from one thread and thread_suspend from another. (I believe the suspend
call is attempting to suspend the forking thread)

We call thread_suspend on all managed threads when doing a GC. We can
avoid the collision of calls if we use mono_gc_invoke_with_gc_lock
to protect the fork call.

I do not believe this is needed on other platforms or chips. I suspect
this is a Rosetta problem that we need to work around.