Do not pass mrgctx for default interface instance methods
This is a followup fix from a previous PR
https://github.com/Unity-Technologies/mono/pull/2152

The repo in the original case was dealing with static methods not having
mrgctx set at the call site.

This commit address a new bug introduced by #2152 where the mrgctx is
being set for an instance method and that register has data in it
that is being clobbered.