Fix Incorrect Xml Read Deserializing a Byte Array
Change from using ReadString() to ReadElementString() when
deserialization a byte array.  The expectation ToByteArrayHex() and
ToByteArrayBase64() is that the XmlDocument will be left at the start of
next Xml element.  But with just a ReadString() the document is left at
the end tag of the current element, not the start of the next element.
Changing to ReadElementString() changes this behavior.